Migrating to Bevy `0.13` (#48)
Bevy `0.13` brought many breaking changes some of which affected us:

1. Texture atlas has been split into a resource and a component, and texture sprite component has been removed in favor of a unified `Sprite` component;
2. For `@fragment` shaders, binding groups are now indexed from `1` rather than from `0`;
3. Renaming of input types;
4. UI panics if there's ever a cameraless frame. Fixed by inserting a `'static` camera component with `is_active: false`;
5. and more not worth mentioning.

The whole migration guide can be found [here](https://bevyengine.org/learn/migration-guides/0-12-to-0-13/)

+++ bevy_magic_light_2d/src/gi/resource.rs
@@ -86,8 +86,8 @@ pub struct ComputedTargetSizes {
 impl ComputedTargetSizes {
     pub fn from_window(window: &Window, params: &TargetScalingParams) -> Self {
         let primary_size = Vec2::new(
-            (window.physical_width() as f64 / window.scale_factor()) as f32,
-            (window.physical_height() as f64 / window.scale_factor()) as f32,
+            window.physical_width() as f32 / window.scale_factor(),
+            window.physical_height() as f32 / window.scale_factor(),
         );
 
         let mut sizes = Self::default();
